tests/wpt/web-platform-tests/scroll-to-text-fragment/scroll-to-text-fragment.html

<!doctype html>
<title>Navigating to a text fragment anchor</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
let test_cases = [
  { fragment: '#', expect_scroll: false },
  { fragment: '##targetText=test', expect_scroll: true },
  { fragment: '##targetText=this,page', expect_scroll: true },
  { fragment: '##targetText=this-,is,test', expect_scroll: true },
  { fragment: '##targetText=this-,is,test,-page', expect_scroll: true },
  { fragment: '##targetText=this-,is,page,-none', expect_scroll: false },
  { fragment: '##targetText=this,test,-page', expect_scroll: true },
  { fragment: '##targetText=this%20is%20a%20test%20page', expect_scroll: true },
  { fragment: '##targetText=this&targetText=test,page', expect_scroll: true },
  { fragment: '#pagestate##targetText=test', expect_scroll: true },
  { fragment: '#pagestate##targetText=nomatch', expect_scroll: false },
];

for (const test_case of test_cases) {
  promise_test(t => new Promise(resolve => {
    let channel = new BroadcastChannel('scroll-to-text-fragment');
    channel.addEventListener("message", e => {
      resolve(e.data.didScrollToTarget);
    }, {once: true});

    test_driver.bless('Open a URL with a text fragment anchor', () => {
      window.open('scroll-to-text-fragment-target.html' + test_case.fragment, '_blank', 'noopener');
    });
  }).then(scroll => {
    assert_equals(scroll, test_case.expect_scroll,
                  'Expected ' + test_case.fragment + (test_case.expect_scroll ? ' to scroll.' : ' to not scroll.'));
  }), 'Test navigation with text fragment anchor ' + test_case.fragment);
}
</script>
